Create SystemCommandExecutor Interface
