+++ molgenis-compute-core/src/main/java/org/molgenis/compute5/sysexecutor/SystemCommandExecutor.java
@@ -1,0 +1,14 @@
+package org.molgenis.compute5.sysexecutor;
+
+public interface SystemCommandExecutor
+{
+	public void setWorkingDirectory(String workingDirectory);
+	
+	public void setEnvironmentVar(String name, String value);
+	 
+	public String getCommandOutput();
+	 
+	public String getCommandError();
+	 
+	public int runCommand(String commandLine) throws Exception;
+}
